Tests/LibWeb/Text/expected/URL/url-search-params-iterator-iterator.txt

[object Iterator]
[object Iterator]
true
hello
1
[object Window]
